chore(relayer): refactor event sources
