+++ omni-relayer/src/startup/event_handlers.rs
@@ -1,20 +1,16 @@
 use std::str::FromStr;
-use std::sync::Arc;
 
 use alloy::primitives::{Address, TxHash};
 use anyhow::{Context, Result};
-use async_nats::jetstream::consumer::PullConsumer;
 use bridge_indexer_types::documents_types::{
-    OmniEvent, OmniEventData, OmniMetaEvent, OmniMetaEventDetails, OmniTransactionEvent,
-    OmniTransactionOrigin, OmniTransferMessage,
+    OmniMetaEvent, OmniMetaEventDetails, OmniTransactionEvent, OmniTransactionOrigin,
+    OmniTransferMessage,
 };
-use mongodb::{Client, Collection, change_stream::event::ResumeToken, options::ClientOptions};
 use omni_types::{
     ChainKind, Fee, OmniAddress, TransferId, TransferIdKind, UnifiedTransferId,
     near_events::OmniBridgeEvent,
 };
 use solana_sdk::pubkey::Pubkey;
-use tokio_stream::StreamExt;
 use tracing::{debug, info, warn};
 
 use crate::{
@@ -23,7 +19,7 @@ use crate::{
     workers::{self, RetryableEvent},
 };
 
-const OMNI_EVENTS: &str = "omni_events";
+pub(super) const OMNI_EVENTS: &str = "omni_events";
 
 fn near_event_key(origin_transaction_id: &str, origin_nonce: u64) -> String {
     utils::redis::composite_key(&[origin_transaction_id, &origin_nonce.to_string()])
@@ -153,7 +149,7 @@ fn get_utxo_chain_token(config: &config::Config, chain: ChainKind) -> Option<Omn
 }
 
 #[allow(clippy::too_many_lines)]
-async fn handle_transaction_event(
+pub(super) async fn handle_transaction_event(
     config: &config::Config,
     redis_connection_manager: &mut redis::aio::ConnectionManager,
     nats: Option<&utils::nats::NatsClient>,
@@ -634,7 +630,25 @@ async fn handle_transaction_event(
                     chain: event.transfer_id.origin_chain,
                     btc_tx_hash: utxo_id.tx_hash,
                     vout: utxo_id.vout,
-                    deposit_msg: deposit_msg.clone(),
+                    deposit_msg: crate::types::DepositMsg {
+                        recipient_id: deposit_msg.recipient_id.clone(),
+                        post_actions: deposit_msg.post_actions.clone().map(|actions| {
+                            actions
+                                .into_iter()
+                                .map(|a| crate::types::PostAction {
+                                    receiver_id: a.receiver_id,
+                                    amount: near_sdk::json_types::U128(a.amount.0),
+                                    memo: a.memo,
+                                    msg: a.msg,
+                                    gas: a.gas.map(|g| g.as_gas()),
+                                })
+                                .collect()
+                        }),
+                        extra_msg: deposit_msg.extra_msg.clone(),
+                        safe_deposit: deposit_msg.safe_deposit.clone().map(|sd| {
+                            crate::types::SafeDepositMsg { msg: sd.msg }
+                        }),
+                    },
                 },
             )
             .await;
@@ -681,7 +695,7 @@ async fn handle_transaction_event(
     Ok(())
 }
 
-async fn handle_meta_event(
+pub(super) async fn handle_meta_event(
     config: &config::Config,
     redis_connection_manager: &mut redis::aio::ConnectionManager,
     nats: Option<&utils::nats::NatsClient>,
@@ -791,256 +805,4 @@ async fn handle_meta_event(
     }
 
     Ok(())
-}
-
-async fn watch_omni_events_collection(
-    collection: &Collection<OmniEvent>,
-    config: &config::Config,
-    redis_connection_manager: &mut redis::aio::ConnectionManager,
-    start_timestamp: Option<u32>,
-) -> Result<()> {
-    let mut stream = if let Some(time) = start_timestamp {
-        info!("Starting from timestamp: {time}");
-
-        collection
-            .watch()
-            .start_at_operation_time(mongodb::bson::Timestamp { time, increment: 0 })
-            .await?
-    } else {
-        let resume_token: Option<ResumeToken> = utils::redis::get_last_processed::<&str, String>(
-            config,
-            redis_connection_manager,
-            utils::redis::MONGODB_OMNI_EVENTS_RT,
-        )
-        .await
-        .and_then(|rt| serde_json::from_str(&rt).ok())
-        .unwrap_or_default();
-
-        info!("Resuming from token: {resume_token:?}");
-
-        collection.watch().resume_after(resume_token).await?
-    };
-
-    while let Some(change) = stream.next().await {
-        match change {
-            Ok(doc) => {
-                if let Some(event) = doc.full_document {
-                    match event.event {
-                        OmniEventData::Transaction(transaction_event) => {
-                            tokio::spawn({
-                                let mut redis_connection_manager = redis_connection_manager.clone();
-                                let config = config.clone();
-
-                                async move {
-                                    if let Err(err) = handle_transaction_event(
-                                        &config,
-                                        &mut redis_connection_manager,
-                                        None,
-                                        event.transaction_id,
-                                        transaction_event.transfer_id.clone(),
-                                        event.origin,
-                                        transaction_event,
-                                    )
-                                    .await
-                                    {
-                                        warn!("Failed to handle transaction event: {err:?}");
-                                    }
-                                }
-                            });
-                        }
-                        OmniEventData::Meta(meta_event) => {
-                            if config.bridge_indexer.is_whitelist_active() {
-                                continue;
-                            }
-
-                            tokio::spawn({
-                                let mut redis_connection_manager = redis_connection_manager.clone();
-                                let config = config.clone();
-
-                                async move {
-                                    if let Err(err) = handle_meta_event(
-                                        &config,
-                                        &mut redis_connection_manager,
-                                        None,
-                                        event.transaction_id,
-                                        event.origin,
-                                        meta_event,
-                                    )
-                                    .await
-                                    {
-                                        warn!("Failed to handle meta event: {err:?}");
-                                    }
-                                }
-                            });
-                        }
-                    }
-                }
-            }
-            Err(err) => warn!("Error watching changes: {err:?}"),
-        }
-
-        if let Some(ref resume_token) = stream
-            .resume_token()
-            .and_then(|rt| serde_json::to_string(&rt).ok())
-        {
-            utils::redis::update_last_processed(
-                config,
-                redis_connection_manager,
-                utils::redis::MONGODB_OMNI_EVENTS_RT,
-                resume_token,
-            )
-            .await;
-        }
-    }
-
-    Ok(())
-}
-
-pub async fn start_indexer(
-    config: &config::Config,
-    redis_connection_manager: &mut redis::aio::ConnectionManager,
-    start_timestamp: Option<u32>,
-) -> Result<()> {
-    info!("Connecting to bridge-indexer");
-
-    let Some(ref uri) = config.bridge_indexer.mongodb_uri else {
-        anyhow::bail!("MONGODB_URI is not set");
-    };
-    let Some(ref db_name) = config.bridge_indexer.db_name else {
-        anyhow::bail!("DB_NAME is not set");
-    };
-
-    let client_options = ClientOptions::parse(uri).await?;
-    let client = Client::with_options(client_options)?;
-
-    let db = client.database(db_name);
-    let omni_events_collection = db.collection::<OmniEvent>(OMNI_EVENTS);
-
-    loop {
-        info!("Starting a mongodb stream that track changes in {OMNI_EVENTS}");
-
-        if let Err(err) = watch_omni_events_collection(
-            &omni_events_collection,
-            config,
-            redis_connection_manager,
-            start_timestamp,
-        )
-        .await
-        {
-            warn!("Error watching changes: {err:?}");
-        }
-
-        warn!("Mongodb stream was closed, restarting...");
-        tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
-    }
-}
-
-async fn process_nats_message(
-    config: &config::Config,
-    redis_connection_manager: &mut redis::aio::ConnectionManager,
-    nats: Option<&utils::nats::NatsClient>,
-    event: OmniEvent,
-) -> Result<()> {
-    match event.event {
-        OmniEventData::Transaction(transaction_event) => {
-            handle_transaction_event(
-                config,
-                redis_connection_manager,
-                nats,
-                event.transaction_id,
-                transaction_event.transfer_id.clone(),
-                event.origin,
-                transaction_event,
-            )
-            .await?;
-        }
-        OmniEventData::Meta(meta_event) => {
-            handle_meta_event(
-                config,
-                redis_connection_manager,
-                nats,
-                event.transaction_id,
-                event.origin,
-                meta_event,
-            )
-            .await?;
-        }
-    }
-
-    Ok(())
-}
-
-async fn subscribe_to_omni_events(
-    consumer: &PullConsumer,
-    config: &config::Config,
-    redis_connection_manager: &mut redis::aio::ConnectionManager,
-    nats: Option<&utils::nats::NatsClient>,
-) -> Result<()> {
-    let mut messages = consumer
-        .messages()
-        .await
-        .context("Failed to start consuming NATS messages")?;
-
-    while let Some(msg) = messages.next().await {
-        let msg = msg.context("NATS message error")?;
-
-        let omni_event: OmniEvent = match serde_json::from_slice(&msg.payload) {
-            Ok(event) => event,
-            Err(err) => {
-                warn!("Failed to deserialize OmniEvent from NATS, terminating message: {err:?}");
-                msg.ack_with(async_nats::jetstream::AckKind::Term)
-                    .await
-                    .ok();
-                continue;
-            }
-        };
-
-        if let Err(err) =
-            process_nats_message(config, redis_connection_manager, nats, omni_event).await
-        {
-            warn!("Failed to process NATS message: {err:?}");
-            continue;
-        }
-
-        if let Err(err) = msg.ack().await {
-            warn!("Failed to ack NATS message: {err:?}");
-        }
-    }
-
-    Ok(())
-}
-
-pub async fn start_indexer_nats(
-    config: &config::Config,
-    redis_connection_manager: &mut redis::aio::ConnectionManager,
-    nats_client: Arc<utils::nats::NatsClient>,
-) -> Result<()> {
-    let nats_config = config.nats.as_ref().context("NATS config is not set")?;
-
-    loop {
-        info!("Starting NATS subscription for OmniEvents");
-
-        let consumer = match nats_client.omni_consumer(nats_config).await {
-            Ok(consumer) => consumer,
-            Err(err) => {
-                warn!("Failed to create NATS consumer: {err:?}");
-                tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
-                continue;
-            }
-        };
-
-        if let Err(err) = subscribe_to_omni_events(
-            &consumer,
-            config,
-            redis_connection_manager,
-            Some(nats_client.as_ref()),
-        )
-        .await
-        {
-            warn!("Error in NATS subscription: {err:?}");
-        }
-
-        warn!("NATS subscription ended, restarting...");
-        tokio::time::sleep(tokio::time::Duration::from_secs(1)).await;
-    }
 }
