+++ omni-relayer/src/startup/native_indexers/mod.rs
@@ -1,0 +1,3 @@
+pub mod evm;
+pub mod near;
+pub mod solana;
